PI-317 Separate auditing functionality into a separate library
The risk-assessment-scores-to-delius service doesn't need to audit, and doesn't even need a user record in the Delius database - so it makes sense to move these components into a separate, optional library.
