+++ libs/audit/src/main/kotlin/uk/gov/justice/digital/hmpps/audit/service/AuditedInteractionService.kt
@@ -9,7 +9,7 @@ import uk.gov.justice.digital.hmpps.audit.BusinessInteractionNotFoundException
 import uk.gov.justice.digital.hmpps.audit.InteractionCode
 import uk.gov.justice.digital.hmpps.audit.repository.AuditedInteractionRepository
 import uk.gov.justice.digital.hmpps.audit.repository.BusinessInteractionRepository
-import uk.gov.justice.digital.hmpps.config.security.ServicePrincipal
+import uk.gov.justice.digital.hmpps.security.ServicePrincipal
 
 @Service
 class AuditedInteractionService(
