+++ libs/audit/src/main/kotlin/uk/gov/justice/digital/hmpps/datasource/OptimisationContext.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.config.datasource
+package uk.gov.justice.digital.hmpps.datasource
 
 object OptimisationContext {
     val offenderId: ThreadLocal<Long> = ThreadLocal()
