+++ libs/audit/src/main/kotlin/uk/gov/justice/digital/hmpps/security/ServiceContext.kt
@@ -1,8 +1,7 @@
-package uk.gov.justice.digital.hmpps.config
+package uk.gov.justice.digital.hmpps.security
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.config.security.ServicePrincipal
 import uk.gov.justice.digital.hmpps.user.UserService
 import javax.annotation.PostConstruct
