+++ libs/audit/src/test/kotlin/uk/gov/justice/digital/hmpps/datasource/DeliusConnectionProviderTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.config
+package uk.gov.justice.digital.hmpps.datasource
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -7,7 +7,7 @@ import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import uk.gov.justice.digital.hmpps.config.datasource.DeliusConnectionProvider
+import uk.gov.justice.digital.hmpps.security.ServiceContext
 import uk.gov.justice.digital.hmpps.user.User
 import uk.gov.justice.digital.hmpps.user.UserService
 import java.sql.Connection
