+++ libs/commons/build.gradle.kts
@@ -2,14 +2,11 @@ import uk.gov.justice.digital.hmpps.extensions.ClassPathExtension
 
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-security")
-    implementation("org.springframework.boot:spring-boot-starter-oauth2-client")
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("org.springframework:spring-jms")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
     implementation(libs.amazon.sqs)
     implementation(libs.insights)
-    implementation(libs.openfeign)
     implementation(libs.sentry)
 
     testImplementation("org.springframework.boot:spring-boot-starter-test")
@@ -18,13 +15,7 @@ dependencies {
 
 configure<ClassPathExtension> {
     jacocoExclusions = listOf(
-        "**/AuditingConfig*",
-        "**/ContextRunnable*",
-        "**/ConnectionProviderConfig*",
         "**/exception/**",
-        "**/JmsConfig*",
-        "**/OracleCondition*",
-        "**/SecurityConfig*",
-        "**/ThreadConfig*",
+        "**/config/**",
     )
 }
