+++ projects/pre-sentence-reports-to-delius/build.gradle.kts
@@ -3,6 +3,7 @@ import uk.gov.justice.digital.hmpps.extensions.ClassPathExtension
 apply(plugin = "com.google.cloud.tools.jib")
 
 dependencies {
+    implementation(project(":libs:audit"))
     implementation(project(":libs:commons"))
     implementation(project(":libs:oauth-client"))
