+++ projects/risk-assessment-scores-to-delius/deploy/database/access.yml
@@ -2,8 +2,3 @@ database:
   access:
     username_key: /risk-assessment-scores-to-delius/db-username
     password_key: /risk-assessment-scores-to-delius/db-password
-
-  audit:
-    username: RiskAssessmentScoresToDelius
-    forename: Risk Assessment Scores
-    surname: Service
